Using iframe communitcation to send data to preview (reduction in network usage)
Previously code was put in the query params which were sent to the server. This meant that every change invalidated the cache (This could have been overwritten but nevertheless the query would have still been sent by the browser to the server). This is a big step to improvements on bad internet connections